+++ lti_app/tests/test_views.py
@@ -8,16 +8,17 @@ from django.shortcuts import reverse
 from django.test import Client, RequestFactory, TestCase
 from wimsapi import Class, Sheet, User
 
-from api.models import LMS, WIMS, WimsClass
 from lti_app import views
 # URL to the WIMS server used for tests, the server must recogned ident 'myself' and passwd 'toto'
 from lti_app.enums import Role
+from lti_app.models import LMS, WIMS, WimsClass
 
 
 WIMS_URL = os.getenv("WIMS_URL") or "http://localhost:7777/wims/wims.cgi"
 
 KEY = 'provider1'
 SECRET = 'secret1'
+PASSWORD = "password"
 
 
 
@@ -609,8 +610,8 @@ class WimsActivityTestCase(TestCase):
         
         self.assertIn(WIMS_URL, r.url)
         self.assertIn("sh=1", r.url)
-
-
+    
+    
     def test_wims_activity_ok_overwrite_sourcedid(self):
         params = {
             'lti_message_type':                   'basic-lti-launch-request',
@@ -633,17 +634,17 @@ class WimsActivityTestCase(TestCase):
             'oauth_nonce':                        oauth2.generate_nonce(),
             'roles':                              Role.LEARNER.value,
         }
-
+        
         norm_params = oauth_signature.normalize_parameters([(k, v) for k, v in params.items()])
-
+        
         uri = oauth_signature.normalize_base_string_uri(
             "https://testserver" + reverse("lti:wims_activity", args=[1, 1]))
         base_string = oauth_signature.construct_base_string("POST", uri, norm_params)
-
+        
         params['oauth_signature'] = oauth_signature.sign_hmac_sha1(base_string, SECRET, None)
         request = RequestFactory().post(reverse("lti:wims_activity", args=[1, 1]), secure=True)
         request.POST = params
-
+        
         wims = WIMS.objects.create(url=WIMS_URL, name="WIMS UPEM", ident="myself",
                                    passwd="toto",
                                    rclass="myclass")
@@ -656,9 +657,101 @@ class WimsActivityTestCase(TestCase):
         WimsClass.objects.create(lms=lms, lms_uuid="77777", wims=wims, qclass=wclass.qclass,
                                  name="test1")
         wclass.additem(Sheet("Titre", "Description", sheetmode=1))
-
+        
         views.wims_activity(request, 1, 1)
         r = views.wims_activity(request, 1, 1)
-
+        
         self.assertIn(WIMS_URL, r.url)
         self.assertIn("sh=1", r.url)
+
+
+
+class BaseLinksViewTestCase(TestCase):
+    
+    @classmethod
+    def setUpTestData(cls):
+        cls.client = Client()
+        cls.lms1 = LMS.objects.create(uuid="elearning.upem.fr", url="https://elearning.u-pem.fr/",
+                                      name="No WIMS", key="provider1", secret="secret1")
+        cls.lms2 = LMS.objects.create(uuid="elearning.test.fr", url="https://elearning.test.fr/",
+                                      name="One WIMS", key="provider2", secret="secret1")
+        cls.lms3 = LMS.objects.create(uuid="elearning.test.fr", url="https://elearning.test.fr/",
+                                      name="Two WIMS", key="provider3", secret="secret1")
+        
+        cls.wims1 = WIMS.objects.create(url="www.lti_app.com", name="One", ident="myself",
+                                        passwd="toto", rclass="myclass")
+        cls.wims2 = WIMS.objects.create(url=WIMS_URL, name="Two", ident="myself", passwd="toto",
+                                        rclass="myclass")
+        
+        cls.wims1.allowed_lms.add(cls.lms2)
+        cls.wims1.allowed_lms.add(cls.lms3)
+        cls.wims2.allowed_lms.add(cls.lms3)
+        
+        c1 = Class("myclass", "Classe 1", "Institution", "email", PASSWORD,
+                   User("supervisor", "First", "Last", PASSWORD))
+        c2 = Class("myclass", "Classe 2", "Institution", "email", PASSWORD,
+                   User("supervisor", "First", "Last", PASSWORD))
+        c1.save(cls.wims2.url, cls.wims2.ident, cls.wims2.passwd)
+        c2.save(cls.wims2.url, cls.wims2.ident, cls.wims2.passwd)
+        cls.class1 = WimsClass.objects.create(lms=cls.lms3, wims=cls.wims2, lms_uuid=1,
+                                              qclass=c1.qclass, name=c1.name)
+        cls.class2 = WimsClass.objects.create(lms=cls.lms3, wims=cls.wims2, lms_uuid=2,
+                                              qclass=c2.qclass, name=c2.name)
+        
+        cls.sheet1 = Sheet("Sheet1", "Desc1")
+        cls.sheet2 = Sheet("Sheet2", "Desc2")
+        c2.additem(cls.sheet1)
+        c2.additem(cls.sheet2)
+
+
+
+class LMSTestCase(BaseLinksViewTestCase):
+    
+    def test_lms(self):
+        response = self.client.get(reverse("lti:lms"))
+        self.assertTemplateUsed(response, 'lti_app/base.html')
+        self.assertTemplateUsed(response, 'lti_app/lms.html')
+        self.assertContains(response, 'No WIMS')
+        self.assertContains(response, 'One WIMS')
+        self.assertContains(response, 'Two WIMS')
+
+
+
+class WIMSTestCase(BaseLinksViewTestCase):
+    
+    def test_wims(self):
+        response = self.client.get(reverse("lti:wims", args=[self.lms3.pk]))
+        self.assertTemplateUsed(response, 'lti_app/base.html')
+        self.assertTemplateUsed(response, 'lti_app/wims.html')
+        self.assertContains(response, 'One')
+        self.assertContains(response, 'Two')
+        self.assertContains(response, reverse('lti:wims_class', args=[self.wims1.pk]))
+
+
+
+class ClassesTestCase(BaseLinksViewTestCase):
+    
+    def test_classes(self):
+        response = self.client.get(reverse("lti:classes", args=[self.lms3.pk, self.wims2.pk]))
+        self.assertTemplateUsed(response, 'lti_app/base.html')
+        self.assertTemplateUsed(response, 'lti_app/classes.html')
+        self.assertContains(response, 'Classe 1')
+        self.assertContains(response, 'Classe 2')
+
+
+
+class SheetTestCase(BaseLinksViewTestCase):
+    
+    def test_sheets(self):
+        response = self.client.post(
+            reverse("lti:sheets", args=[self.lms3.pk, self.wims2.pk, self.class2.pk]),
+            {"password": PASSWORD}
+        )
+        self.assertContains(response, 'Sheet1')
+        self.assertContains(response, 'Sheet2')
+        self.assertContains(response,
+                            reverse('lti:wims_activity', args=[self.wims2.pk, self.sheet1.qsheet]))
+        self.assertContains(response,
+                            reverse('lti:wims_activity', args=[self.wims2.pk, self.sheet2.qsheet]))
+        self.assertTemplateUsed(response, 'lti_app/base.html')
+        self.assertTemplateUsed(response, 'lti_app/activities.html')
