+++ lti_app/urls.py
@@ -15,7 +15,10 @@ from lti_app import views
 app_name = "lti"
 
 urlpatterns = [
-    path('<int:wims_pk>/', views.wims_class, name="wims_class"),
-    path('<int:wims_pk>/<int:activity_pk>/', views.wims_activity, name="wims_activity"),
-    path('links/', views.links, name="links"),
+    path('lti/<int:wims_pk>/', views.wims_class, name="wims_class"),
+    path('lti/<int:wims_pk>/<int:activity_pk>/', views.wims_activity, name="wims_activity"),
+    path('', views.lms, name="lms"),
+    path('<int:lms_pk>/', views.wims, name="wims"),
+    path('<int:lms_pk>/<int:wims_pk>/', views.classes, name="classes"),
+    path('<int:lms_pk>/<int:wims_pk>/<int:wclass_pk>/', views.activities, name="sheets"),
 ]
