+++ wimsLTI/config.py
@@ -27,4 +27,4 @@ For more informations, see:
 
 # THE LINE BELOW MUST BE UNCOMMENTED BEFORE DEPLOYEMENT
 # DEBUG = False
-ALLOWED_HOSTS = ['pl-test3.u-pem.fr']
+ALLOWED_HOSTS = ['pl-test3.u-pem.fr', '127.0.0.1']
