+++ wimsLTI/urls.py
@@ -12,6 +12,5 @@ from django.urls import include, path
 
 urlpatterns = [
     path('admin/', admin.site.urls),
-    path('api/', include('api.urls', namespace='api')),
-    path('lti/', include('lti_app.urls', namespace='lti')),
+    path('', include('lti_app.urls', namespace='lti')),
 ]
